split problem data into different header files
