+++ traj_opt/velocity_partials.h
@@ -1,0 +1,41 @@
+#pragma once
+
+#include <vector>
+
+#include "drake/common/eigen_types.h"
+
+namespace drake {
+namespace traj_opt {
+
+using Eigen::MatrixXd;
+
+/**
+ * Struct storing gradients of generalized velocities (v) with respect to
+ * generalized positions (q).
+ *
+ * TODO(vincekurtz): extend to quaternion DoFs, where these quantities are
+ * different for each timestep, and include a factor of N+(q).
+ */
+struct VelocityPartials {
+  VelocityPartials(const int num_steps, const int nv, const int nq) {
+    dvt_dqt.assign(num_steps + 1, MatrixXd(nv, nq));
+    dvt_dqm.assign(num_steps + 1, MatrixXd(nv, nq));
+
+    // Derivatives w.r.t. q(-1) are undefined
+    dvt_dqm[0].setConstant(nv, nq, NAN);
+  }
+  // Partials of v_t w.r.t. q_t at each time step:
+  //
+  //    [d(v_0)/d(q_0), d(v_1)/d(q_1), ... , d(v_{num_steps})/d(q_{num_steps}) ]
+  //
+  std::vector<MatrixXd> dvt_dqt;
+
+  // Partials of v_t w.r.t. q_{t-1} at each time step:
+  //
+  //    [NaN, d(v_1)/d(q_0), ... , d(v_{num_steps})/d(q_{num_steps-1}) ]
+  //
+  std::vector<MatrixXd> dvt_dqm;
+};
+
+}  // namespace traj_opt
+}  // namespace drake
